janitor: move to under koding/workers
