+++ go/src/koding/workers/janitor/result.go
