+++ go/src/koding/workers/janitor/secretkey/secretkey.go
